fix(wardend): do the node key sync only if it can be found

+++ charts/wardend/scripts/init.sh
@@ -21,15 +21,17 @@ DST_NOKE_KEY="$WARDEND_HOME/config/node_key.json"
 SRC_HASH=$(sha256sum "$SRC_NODE_KEY" | awk '{ print $1 }')
 DST_HASH=$(sha256sum "$DST_NOKE_KEY" | awk '{ print $1 }')
 
-if [ "$SRC_HASH" != "$DST_HASH" ]; then
-  echo "Node key has changed. Copying new version..."
-  cp "$SRC_NOKE_KEY" "$DST_NOKE_KEY"
-  if [ $? -eq 0 ]; then
-      echo "Node key updated successfully."
+if [ -e "$SRC_NOKE_KEY" ]; then
+  if [ "$SRC_HASH" != "$DST_HASH" ]; then
+    echo "Node key has changed. Copying new version..."
+    cp "$SRC_NOKE_KEY" "$DST_NOKE_KEY"
+    if [ $? -eq 0 ]; then
+        echo "Node key updated successfully."
+    else
+        echo "Failed to update node_key"
+        exit 1
+    fi
   else
-      echo "Failed to update node_key"
-      exit 1
+      echo "Node Key is up to date. No action needed."
   fi
-else
-    echo "Node Key is up to date. No action needed."
 fi
